# specs/003-fastapi-better-auth/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI Better-Auth Backend API
  description: |
    Production-ready API with Better-Auth integration for authentication.
    Authentication is handled by Better-Auth (separate service), this API validates JWTs.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.example.com
    description: Production
...
paths:
...
  /api/v1/auth/session:
    get:
      tags: [auth]
      summary: Get current session
      description: Validates JWT and returns current user session
      operationId: getSession
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user profile
      description: Returns the authenticated user's profile
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # User Management Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better-Auth
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # User Schemas
    User:
      type: object
      required: [id, email, email_verified, role, created_at, updated_at]
      properties:
        id:
          type: string
          example: "01HX..."
        email:
          type: string
          format: email
          example: "user@example.com"
        email_verified:
          type: boolean
          example: true
        name:
          type: string
          nullable: true
          example: "John Doe"
        image:
          type: string
          nullable: true
          format: uri
        role:
          type: string
          enum: [admin, user]
          example: user
        banned:
          type: boolean
          example: false
        ban_reason:
          type: string
          nullable: true
        ban_expires:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
...
    Session:
      type: object
      required: [id, user_id, expires_at, created_at]
      properties:
        id:
          type: string
        user_id:
          type: string
        expires_at:
          type: string
          format: date-time
        ip_address:
          type: string
          nullable: true
        user_agent:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    SessionResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            session:
              $ref: '#/components/schemas/Session'
            user:
              $ref: '#/components/schemas/User'
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
...